+++ src/channel.rs
@@ -4,12 +4,14 @@ use std::borrow::Cow;
 use std::collections::hash_map::Entry;
 use std::collections::HashMap;
 use std::ffi::{CStr, CString};
+use std::future::Future;
 use std::sync::Arc;
 use std::time::Duration;
 use std::{cmp, i32, ptr};
 
-use crate::grpc_sys::{
-    self, gpr_timespec, grpc_arg_pointer_vtable, grpc_channel, grpc_channel_args,
+use crate::{
+    grpc_sys::{self, gpr_timespec, grpc_arg_pointer_vtable, grpc_channel, grpc_channel_args},
+    Deadline,
 };
 use libc::{self, c_char, c_int};
 
@@ -17,6 +19,7 @@ use crate::call::{Call, Method};
 use crate::cq::CompletionQueue;
 use crate::env::Environment;
 use crate::error::Result;
+use crate::task::CallTag;
 use crate::task::Kicker;
 use crate::CallOption;
 use crate::ResourceQuota;
@@ -111,7 +114,7 @@ impl ChannelBuilder {
         self
     }
 
-    /// Set maximum message length that the channel can receive. `usize::MAX` means unlimited.
+    /// Set maximum message length that the channel can receive. `-1` means unlimited.
     pub fn max_receive_message_len(mut self, len: i32) -> ChannelBuilder {
         self.options.insert(
             Cow::Borrowed(grpcio_sys::GRPC_ARG_MAX_RECEIVE_MESSAGE_LENGTH),
@@ -442,6 +445,28 @@ impl ChannelBuilder {
         let channel =
             unsafe { grpc_sys::grpc_insecure_channel_create(addr_ptr, args.args, ptr::null_mut()) };
 
+        unsafe { Channel::new(self.env.pick_cq(), self.env, channel) }
+    }
+
+    /// Build an insecure [`Channel`] taking over an established connection from
+    /// a file descriptor. The target string given is purely informative to
+    /// describe the endpoint of the connection. Takes ownership of the given
+    /// file descriptor and will close it when the connection is closed.
+    ///
+    /// This function is available on posix systems only.
+    ///
+    /// # Safety
+    ///
+    /// The file descriptor must correspond to a connected stream socket. After
+    /// this call, the socket must not be accessed (read / written / closed)
+    /// by other code.
+    #[cfg(unix)]
+    pub unsafe fn connect_from_fd(mut self, target: &str, fd: ::std::os::raw::c_int) -> Channel {
+        let args = self.prepare_connect_args();
+        let target = CString::new(target).unwrap();
+        let target_ptr = target.as_ptr();
+        let channel = grpc_sys::grpc_insecure_channel_create_from_fd(target_ptr, fd, args.args);
+
         Channel::new(self.env.pick_cq(), self.env, channel)
     }
 }
@@ -489,7 +514,7 @@ mod secure_channel {
                 )
             };
 
-            Channel::new(self.env.pick_cq(), self.env, channel)
+            unsafe { Channel::new(self.env.pick_cq(), self.env, channel) }
         }
     }
 }
@@ -548,17 +573,83 @@ unsafe impl Send for Channel {}
 unsafe impl Sync for Channel {}
 
 impl Channel {
-    fn new(cq: CompletionQueue, env: Arc<Environment>, channel: *mut grpc_channel) -> Channel {
+    /// Create a new channel. Avoid using this directly and use
+    /// [`ChannelBuilder`] to build a [`Channel`] instead.
+    ///
+    /// # Safety
+    ///
+    /// The given grpc_channel must correspond to an instantiated grpc core
+    /// channel. Takes exclusive ownership of the channel and will close it after
+    /// use.
+    pub unsafe fn new(
+        cq: CompletionQueue,
+        env: Arc<Environment>,
+        channel: *mut grpc_channel,
+    ) -> Channel {
         Channel {
             inner: Arc::new(ChannelInner { _env: env, channel }),
             cq,
         }
     }
 
-    // If try_to_connect is true, the channel will try to establish a connection, potentially
-    // changing the state.
+    /// If try_to_connect is true, the channel will try to establish a connection, potentially
+    /// changing the state.
     pub fn check_connectivity_state(&self, try_to_connect: bool) -> ConnectivityState {
         self.inner.check_connectivity_state(try_to_connect)
+    }
+
+    /// Blocking wait for channel state change or deadline expiration.
+    ///
+    /// `check_connectivity_state` needs to be called to get the current state. Returns false
+    /// means deadline excceeds before observing any state changes.
+    pub fn wait_for_state_change(
+        &self,
+        last_observed: ConnectivityState,
+        deadline: impl Into<Deadline>,
+    ) -> impl Future<Output = bool> {
+        let (cq_f, prom) = CallTag::action_pair();
+        let prom_box = Box::new(prom);
+        let tag = Box::into_raw(prom_box);
+        let should_wait = if let Ok(cq_ref) = self.cq.borrow() {
+            unsafe {
+                grpcio_sys::grpc_channel_watch_connectivity_state(
+                    self.inner.channel,
+                    last_observed,
+                    deadline.into().spec(),
+                    cq_ref.as_ptr(),
+                    tag as *mut _,
+                )
+            }
+            true
+        } else {
+            // It's already shutdown.
+            false
+        };
+        async move { should_wait && cq_f.await.unwrap() }
+    }
+
+    /// Wait for this channel to be connected.
+    ///
+    /// Returns false means deadline excceeds before connection is connected.
+    pub async fn wait_for_connected(&self, deadline: impl Into<Deadline>) -> bool {
+        // Fast path, it's probably connected.
+        let mut state = self.check_connectivity_state(true);
+        if ConnectivityState::GRPC_CHANNEL_READY == state {
+            return true;
+        }
+        let deadline = deadline.into();
+        loop {
+            if self.wait_for_state_change(state, deadline).await {
+                state = self.check_connectivity_state(true);
+                match state {
+                    ConnectivityState::GRPC_CHANNEL_READY => return true,
+                    ConnectivityState::GRPC_CHANNEL_SHUTDOWN => return false,
+                    _ => (),
+                }
+                continue;
+            }
+            return false;
+        }
     }
 
     /// Create a Kicker.