+++ tests-and-examples/tests/cases/mod.rs
@@ -3,7 +3,6 @@
 mod auth_context;
 mod cancel;
 mod credential;
-mod health_check;
 mod kick;
 mod metadata;
 mod misc;